third_party/blink/web_tests/wpt_internal/bluetooth/server/getPrimaryService/gen-delayed-discovery-service-found.https.html

<!-- Generated by //third_party/blink/web_tests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script src="/common/gc.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => gatt.getPrimaryService('heart_rate'))
    .then(services => {
      services = [].concat(services);
      for (let service of services)
        assert_equals(service.uuid, heart_rate.uuid);
    });
}, 'Request for service. Must return even when the services are not ' +
   'immediately discovered');

</script>
